refactor(clawdstrike): extract engine tests into sibling tests/ module
Move the inline `#[cfg(test)] mod tests` out of engine.rs into a
`engine/tests/` subtree split on the existing banners into four themed
parts (core, origin enclave, posture integration, cross-origin isolation).
Shared fixtures (TestExtraGuard, TestAsyncAllowGuard, and the origin/
enclave/posture helper fns) live in tests/mod.rs as pub(super) so each
part can reach them via `use super::*`.

Pure relocation: no test logic changed, all 70 engine tests still run and
pass. First step of the engine.rs god-file decomposition (step 6).
